// Angular/03.DI_Intro-to-RxJS_Services/ServicesRxJS/my-app-playground/src/app/types/user.ts
export interface User {
    name: string;
    age: number;
}

export interface Geo {
    lat: string;
    lng: string;
  };

  export interface Address {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: Geo;
  };

  export interface Company {
    name: string;
    catchPhrase: string;
    bs: string;
  };

  export interface User2 {
    id: number;
    name: string;
    username: string;
    email: string;
    phone: string;
  }

export interface JsonPlaceholderUser extends User2 {
        address: Address;
        website: string;
        company: Company;
};